2024.07.12 14:40

+++ scripts/common/feature.sh
@@ -12,10 +12,10 @@ pause()
 }
 
 # 获取命令序号
-get_user_index()
+input_user_index()
 {
 	#stty erase ^H
-	read value
+	read -r value
 	
 	echo "$value"|[ -n "`sed -n '/^[0-9][0-9]*$/p'`" ]
 	
@@ -26,6 +26,30 @@ get_user_index()
 	fi
 	
 	echo "$result"
+}
+
+# 获取用户选择是否
+input_prompt_confirm()
+{
+	local prompt="$1"
+	local input
+	
+	while true; do
+		printf "\033[1;33m%s\033[0m" "${prompt} (y/n):"
+        read -r input
+		
+		case "${input}" in
+			[Yy])
+				return 0
+				;;
+			[Nn])
+				return 1
+				;;
+			 *)
+				echo "无效输入，请输入 y 或 n."
+				;;
+		esac
+	done
 }
 
 #  打印日志信息
@@ -283,66 +307,80 @@ enum_struct_field()
 # 克隆仓库内容
 clone_repo_contents() 
 {
-	local remote_repo=$1        
-	local package_path_rel=$2
+	# 远程仓库URL
+	local remote_repo=$1
+	
+	# 本地指定路径
+	local local_path=$2
+	
+	# 代理命令
 	local proxy_cmd=$3
 	
-	# 获取?的前缀和后缀字符
-	mark_prefix="${remote_repo%%\?*}"
-	mark_suffix="${remote_repo#*\?}"
+	# 获取.git前缀和后缀字符
+	git_prefix="${remote_repo%%.git*}"
+	git_suffix="${remote_repo#*.git}"
 	
-	if [ -z "${mark_prefix}" ] || [ -z "${mark_suffix}" ]; then
+	if [ -z "${git_prefix}" ] || [ -z "${git_suffix}" ]; then
 		return
 	fi
 	
-	# 获取.git的前缀
-	git_prefix="${mark_prefix%%.git*}"
-	if [ -z "${git_prefix}" ]; then
-		return
-	fi
+	# 获取?前缀和后缀字符
+	suffix_before_mark="${git_suffix%%\?*}"
+	suffix_after_mark="${git_suffix#*\?}"
 	
-	repo_url="${mark_prefix}"	# 远程仓库URL
-	repo_branch=$(echo ${mark_suffix} | awk -F '=' '{print $2; exit}')		# 分支名
-	local_dir_name=$(echo ${git_prefix} | awk -F '/' '{print $NF}')			# 目录名
+	# url地址
+	repo_url="${git_prefix}.git"
+
+	# 远程分支名称
+	repo_branch=$(echo ${suffix_after_mark} | awk -F '=' '{print $2; exit}')
 	
 	# 临时目录，用于克隆远程仓库
 	local temp_dir=$(mktemp -d)
 	
-	# 克隆远程仓库到临时目录
-	${proxy_cmd} git clone --depth 1 --branch $repo_branch $repo_url $temp_dir
+	echo "Cloning branch code... ${repo_branch}"
+
+	# 克隆远程仓库到临时目录 ${proxy_cmd}
+	${proxy_cmd} git clone --depth 1 --branch ${repo_branch} ${repo_url} ${temp_dir}
 	
 	if [ $? -eq 0 ]; then
-		local target_path="$package_path_rel/$local_dir_name"
+		local target_path="${local_path}"
 		
-		if [ -d "$target_path" ]; then
-			echo "Removing old files from $target_path."
-			
+		if [ ! -d "${target_path}" ]; then
+			mkdir -p "${target_path}"
+		fi
+		
+		# 判断目标目录是否为空
+		if [ ! -z "$(ls -A ${target_path})" ]; then
 			# 使用:?防止变量为空时删除根目录
 			rm -rf "${target_path:?}"/*  
-		else
-			# 如果目标路径不存在，创建目标路径
-			mkdir -p "$target_path"
 		fi
-		# 创建目标路径
-		mkdir -p "$package_path_rel/$local_dir_name"
-	
+		
+		echo "Copying repo directory to local...."
+		
 		# 复制克隆的内容到目标路径
-		cp -r $temp_dir/* "$package_path_rel/$local_dir_name/"
+		cp -r ${temp_dir}/* "${local_path}"
 	fi
 	
 	# 清理临时目录
-	rm -rf $temp_dir
+	rm -rf ${temp_dir}
 }
 
 # 添加获取远程仓库指定内容
 get_remote_spec_contents()
 {
-	local remote_repo=$1        # 远程仓库URL
-	local remote_alias=$2		# 远程仓库别名
-	local local_path=$3    		# 本地指定路径
-	local proxy_cmd=$4			# 代理命令
+	# 远程仓库URL
+	local remote_repo=$1
 	
-	# 获取.git的前缀和后缀字符
+	# 远程仓库别名
+	local remote_alias=$2
+	
+	# 本地指定路径
+	local local_path=$3
+	
+	# 代理命令
+	local proxy_cmd=$4
+	
+	# 获取.git前缀和后缀字符
 	git_prefix="${remote_repo%%.git*}"
 	git_suffix="${remote_repo#*.git}"
 
@@ -350,7 +388,7 @@ get_remote_spec_contents()
 		return
 	fi
 	
-	# 获取?的前缀和后缀字符
+	# 获取?前缀和后缀字符
 	suffix_before_mark="${git_suffix%%\?*}"	#
 	suffix_after_mark="${git_suffix#*\?}"	#
 
@@ -358,9 +396,14 @@ get_remote_spec_contents()
 		return
 	fi
 	
-	repo_url="${git_prefix}.git"			# url地址
-	repo_path="${suffix_before_mark}"		# 指定路径
-	repo_branch=$(echo ${suffix_after_mark} | awk -F '=' '{print $2; exit}')	# 远程分支名称
+	# url地址
+	repo_url="${git_prefix}.git"
+	
+	# 指定路径
+	repo_path="${suffix_before_mark}"
+	
+	# 远程分支名称
+	repo_branch=$(echo ${suffix_after_mark} | awk -F '=' '{print $2; exit}')
 	
 	# 临时目录，用于克隆远程仓库
 	local temp_dir=$(mktemp -d)
@@ -372,14 +415,15 @@ get_remote_spec_contents()
 	pushd ${temp_dir} > /dev/null	# cd ${temp_dir}
 	
 	# 添加远程仓库
-	echo "Add remote repository: $remote_alias"
-	git remote add $remote_alias $repo_url || true
+	echo "Add remote repository: ${remote_alias}"
+	git remote add ${remote_alias} ${repo_url} || true
 	
 	# 开启Sparse checkout模式
 	git config core.sparsecheckout true
 	
 	# 配置要检出的目录或文件
 	sparse_file=".git/info/sparse-checkout"
+	
 	if [ ! -e "${sparse_file}" ]; then
 		touch "${sparse_file}"
 	fi
@@ -390,28 +434,30 @@ get_remote_spec_contents()
 	# 从远程将目标目录或文件拉取下来
 	${proxy_cmd} git pull ${remote_alias} ${repo_branch}
 
-	# 目标路径
-	local target_path="${local_path}"
-	if [ ! -d "${target_path}" ]; then
-		mkdir -p "${target_path}"
-	fi
-	
-	# 判断目标目录是否为空
-	if [ ! -z "$(ls -A ${target_path})" ]; then
-		rm -rf "${target_path:?}"/*  
-	fi
-	
-	echo "Copying remote repo directory to local...."
-	
-	if [ -e "${temp_dir}/${repo_path}" ]; then
-		cp -rf ${temp_dir}/${repo_path}/* ${target_path}
+	if [ $? -eq 0 ]; then
+		local target_path="${local_path}"
+		
+		if [ ! -d "${target_path}" ]; then
+			mkdir -p "${target_path}"
+		fi
+		
+		# 判断目标目录是否为空
+		if [ ! -z "$(ls -A ${target_path})" ]; then
+			rm -rf "${target_path:?}"/*  
+		fi
+		
+		echo "Copying remote repo directory to local...."
+		
+		if [ -e "${temp_dir}/${repo_path}" ]; then
+			cp -rf ${temp_dir}/${repo_path}/* ${target_path}
+		fi
 	fi
 	
 	# 返回原始目录
     popd > /dev/null
 	
 	# 清理临时目录
-	rm -rf $temp_dir
+	rm -rf ${temp_dir}
 }
 
 # 显示源码目录
